move ramp and scale tests
